refactor: applying the pattern strategy in generate manifest
Applying the design pattern strategy in the generate manifest for
better maintenance and addition of new implementations.
