+++ lib/utils/generateManifest/generateManifest.utils.js
@@ -4,8 +4,11 @@ import ajvErrors from 'ajv-errors';
 import addKeywords from 'ajv-keywords';
 
 import azionConfigSchema from './helpers/schema.js';
-import { requestBehaviors, responseBehaviors } from './helpers/behaviors.js';
 import convertLegacyConfig from './helpers/convertLegacyConfig.js';
+import ManifestContext from './strategy/manifestContext.js';
+import OriginManifestStrategy from './strategy/implementations/originManifestStrategy.js';
+import CacheManifestStrategy from './strategy/implementations/cacheManifestStrategy.js';
+import RulesManifestStrategy from './strategy/implementations/rulesManifestStrategy.js';
 
 /**
  * Validates the provided configuration against a JSON Schema.
@@ -38,181 +41,15 @@ function jsToJson(inputConfig) {
   const config = convertLegacyConfig(inputConfig);
   validateConfig(config);
 
-  const payloadCDN = {
-    origin: [],
-    cache: [],
-    rules: [],
-  };
+  const payloadCDN = {};
 
-  // Helper function to safely evaluate mathematical expressions
-  const evaluateMathExpression = (expression) => {
-    if (typeof expression === 'number') {
-      return expression;
-    }
-    if (/^[0-9+\-*/.() ]+$/.test(expression)) {
-      // eslint-disable-next-line no-eval
-      return eval(expression);
-    }
-    throw new Error(`Expression is not purely mathematical: ${expression}`);
-  };
-
-  // Convert origin settings
-  if (config && config.origin && config.origin.length > 0) {
-    config.origin.forEach((origin) => {
-      if (origin.type !== 'object_storage' && origin.type !== 'single_origin') {
-        throw new Error(
-          `Rule setOrigin originType '${origin.type}' is not supported`,
-        );
-      }
-      const originSetting = {
-        name: origin.name,
-        origin_type: origin.type,
-      };
-
-      if (origin.type === 'object_storage') {
-        originSetting.bucket = origin.bucket;
-        originSetting.prefix = origin.prefix;
-      }
-      if (origin.type === 'single_origin') {
-        originSetting.addresses = origin.addresses?.map((address) => {
-          return { address };
-        });
-        originSetting.host_header = origin.hostHeader;
-      }
-      payloadCDN.origin.push(originSetting);
-    });
-  }
-
-  // Convert cache settings
-  if (config && config.cache && config.cache.length > 0) {
-    config.cache.forEach((cache) => {
-      const maxAgeSecondsBrowser = cache?.browser
-        ? evaluateMathExpression(cache.browser.maxAgeSeconds)
-        : 0;
-      const maxAgeSecondsEdge = cache?.edge
-        ? evaluateMathExpression(cache.edge.maxAgeSeconds)
-        : 60;
-
-      const cacheSetting = {
-        name: cache.name,
-        browser_cache_settings: cache?.browser ? 'override' : 'honor',
-        browser_cache_settings_maximum_ttl: maxAgeSecondsBrowser,
-        cdn_cache_settings: cache?.edge ? 'override' : 'honor',
-        cdn_cache_settings_maximum_ttl: maxAgeSecondsEdge,
-        enable_caching_for_post: cache?.methods?.post || false,
-        enable_caching_for_options: cache?.methods?.options || false,
-        enable_query_string_sort: cache?.queryStringSort || false,
-      };
-
-      if (cache.cacheByQueryString) {
-        cacheSetting.cache_by_query_string =
-          cache.cacheByQueryString.option === 'varies'
-            ? 'all'
-            : cache.cacheByQueryString.option;
-        if (
-          cache.cacheByQueryString.option === 'whitelist' ||
-          cache.cacheByQueryString.option === 'blacklist'
-        ) {
-          cacheSetting.query_string_fields =
-            cache.cacheByQueryString.list || [];
-        } else {
-          cacheSetting.query_string_fields = [];
-        }
-      }
-
-      if (cache.cacheByCookie) {
-        cacheSetting.cache_by_cookie =
-          cache.cacheByCookie.option === 'varies'
-            ? 'all'
-            : cache.cacheByCookie.option;
-        if (
-          cache.cacheByCookie.option === 'whitelist' ||
-          cache.cacheByCookie.option === 'blacklist'
-        ) {
-          cacheSetting.cookie_names = cache.cacheByCookie.list || [];
-        } else {
-          cacheSetting.cookie_names = [];
-        }
-      }
-
-      payloadCDN.cache.push(cacheSetting);
-    });
-  }
-
-  // Helper function to add behaviors to a rule
-  const addBehaviors = (cdnRule, behaviors, behaviorDefinitions) => {
-    if (behaviors && typeof behaviors === 'object') {
-      Object.entries(behaviors).forEach(([key, value]) => {
-        if (behaviorDefinitions[key]) {
-          const transformedBehavior = behaviorDefinitions[key].transform(
-            value,
-            payloadCDN,
-          );
-          if (Array.isArray(transformedBehavior)) {
-            cdnRule.behaviors.push(...transformedBehavior);
-          } else if (transformedBehavior) {
-            cdnRule.behaviors.push(transformedBehavior);
-          }
-        } else {
-          console.warn(`Unknown behavior: ${key}`);
-        }
-      });
-    }
-  };
-
-  // Convert request rules
-  config?.rules?.request?.forEach((rule, index) => {
-    const cdnRule = {
-      name: rule.name,
-      phase: 'request',
-      description: rule.description ?? '',
-      is_active: rule.active !== undefined ? rule.active : true, // Default to true if not provided
-      order: index + 2, // index starts at 2, because the default rule is index 1
-      criteria: [
-        [
-          {
-            variable: `\${${rule.variable ?? 'uri'}}`,
-            operator: 'matches',
-            conditional: 'if',
-            input_value: rule.match,
-          },
-        ],
-      ],
-      behaviors: [],
-    };
-
-    addBehaviors(cdnRule, rule.behavior, requestBehaviors);
-
-    payloadCDN.rules.push(cdnRule);
-  });
-
-  // Convert response rules
-  if (config?.rules?.response) {
-    config.rules.response.forEach((rule, index) => {
-      const cdnRule = {
-        name: rule.name,
-        phase: 'response',
-        description: rule.description ?? '',
-        is_active: rule.active !== undefined ? rule.active : true, // Default to true if not provided
-        order: index + 2, // index starts at 2, because the default rule is index 1
-        criteria: [
-          [
-            {
-              variable: `\${${rule.variable ?? 'uri'}}`,
-              operator: 'matches',
-              conditional: 'if',
-              input_value: rule.match,
-            },
-          ],
-        ],
-        behaviors: [],
-      };
-
-      addBehaviors(cdnRule, rule.behavior, responseBehaviors);
-
-      payloadCDN.rules.push(cdnRule);
-    });
-  }
+  // Manifest Strategy Pattern
+  const manifestContext = new ManifestContext();
+  manifestContext.setStrategy('origin', new OriginManifestStrategy());
+  manifestContext.setStrategy('cache', new CacheManifestStrategy());
+  // Rules must be last to apply to behaviors (origin, cache...)
+  manifestContext.setStrategy('rules', new RulesManifestStrategy());
+  manifestContext.generate(config, payloadCDN);
 
   return payloadCDN;
 }
